Remove old get_lookup_key_field util
